refactor(M2): 設定DRY・dimアサーション・明示的型キャスト
- jvs_aligned.yaml: model.use_precomputed_durations削除、Hydra自動参照に委譲 (B4)
- precompute_dataset.py: GPU/CPUサマリ出力統一 (B5)
- matcha_tts.py forward(): durations dimアサーション追加、3D→2D squeeze対応 (B6)
- TextMelBatchCollate: duration.long()明示キャスト追加 (B7)

+++ matcha/data/text_mel_datamodule.py
@@ -280,7 +280,7 @@ class TextMelBatchCollate:
             filepaths.append(item["filepath"])
             x_texts.append(item["x_text"])
             if item["durations"] is not None:
-                durations[i, : item["durations"].shape[-1]] = item["durations"]
+                durations[i, : item["durations"].shape[-1]] = item["durations"].long()
 
         y_lengths = torch.tensor(y_lengths, dtype=torch.long)
         x_lengths = torch.tensor(x_lengths, dtype=torch.long)
